+++ Libplanet/Crypto/CryptoConfig.cs
@@ -9,6 +9,7 @@ namespace Libplanet.Crypto
     public static class CryptoConfig
     {
         private static ICryptoBackend<SHA256> _cryptoBackend;
+        private static BlsCryptoBackend<SHA256> _consensusCryptoBackend;
 
         /// <summary>
         /// Global cryptography backend to sign and verify messages.
@@ -17,6 +18,12 @@ namespace Libplanet.Crypto
         {
             get => _cryptoBackend ??= new DefaultCryptoBackend<SHA256>();
             set => _cryptoBackend = value;
+        }
+
+        public static BlsCryptoBackend<SHA256> ConsensusCryptoBackend
+        {
+            get => _consensusCryptoBackend ??= new BlsCryptoBackend<SHA256>();
+            set => _consensusCryptoBackend = value;
         }
     }
 }